To do a proper endsession logout, we need to know the loginProvider and id_token. 1. Store it in the session, which has benifits of not having fat cookies 2. Use asp.net's identity to save the tokens away, but we still need to add the loginProvider to the user claims as that is the key need to retrieve the tokens later.
EDGE CASE:
During a create user scenario, the tokens are NOT saved, so I wrote them into a cookie and then added the login_provider and id_token claim to the user.  This is only done for the create user case.  Any subsequent logins will use the asp.net save tokens stuff.

If you can afford to have a backend inmemory session, use that instead as it is more crisp and no weird trickery going on.